+++ src/bosh-alicloud-cpi/action/attach_disk.go
@@ -14,11 +14,12 @@ import (
 type AttachDiskMethod struct {
 	CallContext
 	disks alicloud.DiskManager
+	instances alicloud.InstanceManager
 	registry registry.Client
 }
 
-func NewAttachDiskMethod(cc CallContext, disks alicloud.DiskManager, rc registry.Client) AttachDiskMethod {
-	return AttachDiskMethod{cc, disks, rc}
+func NewAttachDiskMethod(cc CallContext,disks alicloud.DiskManager, instances alicloud.InstanceManager, rc registry.Client) AttachDiskMethod {
+	return AttachDiskMethod{cc, disks, instances, rc}
 }
 
 
@@ -26,6 +27,25 @@ func (a AttachDiskMethod) AttachDisk(vmCID apiv1.VMCID, diskCID apiv1.DiskCID) e
 	instCid := vmCID.AsString()
 	diskCid := diskCID.AsString()
 	device := ""
+
+	if a.Config.Registry.IsEmpty() {
+		err := a.instances.ChangeInstanceStatus(instCid, ecs.Stopped, func(status ecs.InstanceStatus) (bool, error) {
+			switch status {
+			case ecs.Stopped:
+				return true, nil
+			case ecs.Running:
+				return false, a.instances.StopInstance(instCid)
+			case ecs.Stopping:
+				return false, nil
+			default:
+				return false, fmt.Errorf("unexpect %s for StopInstance", status)
+			}
+		})
+
+		if err != nil {
+			return a.WrapError(err, "stop instance failed")
+		}
+	}
 
 	err := a.disks.ChangeDiskStatus(diskCid, ecs.DiskStatusInUse, func(disk *ecs.DiskItemType) (bool, error) {
 		if disk == nil {
@@ -56,6 +76,26 @@ func (a AttachDiskMethod) AttachDisk(vmCID apiv1.VMCID, diskCID apiv1.DiskCID) e
 	if err != nil {
 		return a.WrapErrorf(err, "update registry failed %s %s", diskCid, instCid)
 	}
+
+	if a.Config.Registry.IsEmpty() {
+		err := a.instances.ChangeInstanceStatus(instCid, ecs.Running, func(status ecs.InstanceStatus) (bool, error) {
+			switch status {
+			case ecs.Stopped:
+				return false, a.instances.StartInstance(instCid)
+			case ecs.Starting:
+				return false, nil
+			case ecs.Running:
+				return true, nil
+			default:
+				return false, fmt.Errorf("unexpect %s for StopInstance", status)
+			}
+		})
+
+		if err != nil {
+			return a.WrapError(err, "stop instance failed")
+		}
+	}
+
 	return nil
 }