+++ src/bosh-alicloud-cpi/alicloud/config.go
@@ -14,10 +14,23 @@ import (
 	"github.com/denverdino/aliyungo/common"
 	"time"
 	"github.com/denverdino/aliyungo/slb"
+
+	"github.com/aliyun/aliyun-oss-go-sdk/oss"
+	"strings"
+	"net"
 )
 
+type InnerType string
+
 const (
-	DefaultOpenApiEndpoint = "cn-zhangjiakou.aliyuncs.com"
+	DefaultClassOSSInnerEndpoint = "oss-cn-hangzhou-internal"
+	DefaultVpcOSSInnerEndpoint   = "oss-cn-hangzhou-internal"
+	OSSSuffix                    = "oss-"
+
+	InnerVpc       = InnerType("VPC")
+	InnerClassic   = InnerType("CLASSIC")
+	PingMethod     = "tcp"
+	TimeoutSeconds = 5
 )
 
 type CloudConfigJson struct {
@@ -36,28 +49,29 @@ type Config struct {
 }
 
 const (
-	UseForceStop			= true
+	UseForceStop = true
 
 	WaitTimeout  = time.Duration(180) * time.Second
 	WaitInterval = time.Duration(5) * time.Second
 
-	DefaultEipWaitSeconds	= 120
-	DefaultSlbWeight = 100
+	DefaultEipWaitSeconds = 120
+	DefaultSlbWeight      = 100
 )
 
 type OpenApi struct {
 	RegionId        string  `json:"region_id"`
 	ZoneId			string	`json:"zone_id"`
+	AccessEndpoint	string 	`json:"access_endpoint"`
 	AccessKeyId     string  `json:"access_key_id"`
 	AccessKeySecret string  `json:"access_key_secret"`
 }
 
 type RegistryConfig struct {
-	User     string			`json:"user"`
-	Password string			`json:"password"`
-	Protocol string			`json:"protocol"`
-	Host     string			`json:"address"`
-	Port     json.Number	`json:"port"`
+	User     string      `json:"user"`
+	Password string      `json:"password"`
+	Protocol string      `json:"protocol"`
+	Host     string      `json:"host"`
+	Port     json.Number `json:"port"`
 }
 
 type AgentConfig struct {
@@ -67,8 +81,8 @@ type AgentConfig struct {
 }
 
 type BlobstoreConfig struct {
-	Provider string          		`json:"provider"`
-	Options  map[string]interface{}	`json:"options"`
+	Provider string                 `json:"provider"`
+	Options  map[string]interface{} `json:"options"`
 }
 
 func (c Config) Validate() error {
@@ -87,6 +101,14 @@ func (c Config) Validate() error {
 
 func (a OpenApi) GetRegion() (common.Region) {
 	return common.Region(a.RegionId)
+}
+
+func (a RegistryConfig) IsEmpty() (bool) {
+	if a.Host == "" {
+		return true
+	} else {
+		return false
+	}
 }
 
 func NewConfigFromFile(configFile string, fs boshsys.FileSystem) (Config, error) {
@@ -135,28 +157,91 @@ func (a RegistryConfig) GetEndpoint() (string) {
 }
 
 func (a BlobstoreConfig) AsRegistrySettings() (registry.BlobstoreSettings) {
-	return registry.BlobstoreSettings {
+	return registry.BlobstoreSettings{
 		Provider: a.Provider,
-		Options: a.Options,
+		Options:  a.Options,
 	}
 }
 
 func (c Config) NewEcsClient() (*ecs.Client) {
-	return ecs.NewClientWithRegion(ecs.ECSDefaultEndpoint, c.OpenApi.AccessKeyId, c.OpenApi.AccessKeySecret, common.Region(c.OpenApi.RegionId))
+	// Obsoleted
+	// ep := "https://ecs." + c.OpenApi.GetEndpoint()
+	// return ecs.NewClientWithEndpoint(ep, c.OpenApi.AccessKeyId, c.OpenApi.AccessKeySecret)
+	return ecs.NewClient(c.OpenApi.AccessKeyId, c.OpenApi.AccessKeySecret)
 }
 
 func (c Config) NewSlbClient() (*slb.Client) {
-	return slb.NewClientWithRegion(slb.SLBDefaultEndpoint, c.OpenApi.AccessKeyId, c.OpenApi.AccessKeySecret, common.Region(c.OpenApi.RegionId))
+	// Obsoleted
+	// ep := "https://slb." + c.OpenApi.GetEndpoint()
+	// return slb.NewClientWithEndpoint(ep, c.OpenApi.AccessKeyId, c.OpenApi.AccessKeySecret)
+	return slb.NewClient(c.OpenApi.AccessKeyId, c.OpenApi.AccessKeySecret)
+}
+
+func (c Config) GetRegistryClient(logger boshlog.Logger) (registry.Client) {
+	if !c.Registry.IsEmpty() {
+		return c.GetHttpRegistryClient(logger)
+	} else {
+		return NewRegistryManager(c, logger)
+	}
+}
+
+func (c Config) NewOssClient(inner bool) (*oss.Client) {
+	ossClient, _ := oss.New(c.GetAvailableOSSEndPoint(inner), c.OpenApi.AccessKeyId, c.OpenApi.AccessKeySecret)
+	return ossClient
+}
+
+func (c Config) GetAvailableOSSEndPoint(inner bool) (string) {
+	return "https://" + c.GetOSSEndPoint(inner) + ".aliyuncs.com"
+}
+
+func (c Config) GetOSSEndPoint(inner bool) (string) {
+	timeOut := time.Duration(TimeoutSeconds) * time.Second
+	ep := GetOSSEndPoint(string(c.OpenApi.GetRegion()), "")
+	if !inner {
+		return ep
+	}
+
+	ep = GetOSSEndPoint("", InnerVpc)
+	if _, err := net.DialTimeout(PingMethod, ep, timeOut); err != nil {
+		fmt.Printf("Ping oss inner vpc endpoint %s ok", ep)
+		return ep
+	}
+
+	ep = GetOSSEndPoint("", InnerClassic)
+	if _, err := net.DialTimeout(PingMethod, ep, timeOut); err != nil {
+		fmt.Printf("Ping oss inner ecs endpoint %s ok", ep)
+		return ep
+	}
+
+	ep = GetOSSEndPoint(string(c.OpenApi.GetRegion()), "")
+	return ep
+}
+
+// types allows ["VPC", "CLASSIC"], then return inner endpoint
+// otherwise return endpoint by region
+func GetOSSEndPoint(region string, types InnerType) string {
+	if types == InnerVpc {
+		return DefaultVpcOSSInnerEndpoint
+	}
+
+	if types == InnerClassic {
+		return DefaultClassOSSInnerEndpoint
+	}
+
+	if strings.HasPrefix(region, OSSSuffix) {
+		return region
+	}
+	return OSSSuffix + region
 }
 
 func (c Config) GetHttpRegistryClient(logger boshlog.Logger) (registry.Client) {
 	r := c.Registry
 
 	port, _ := r.Port.Int64()
-	clientOptions := registry.ClientOptions {
+	clientOptions := registry.ClientOptions{
 		Protocol: r.Protocol,
-		Host: r.Host,
-		Port: int(port),
+		Host:     r.Host,
+		Port:     int(port),
 		Username: r.User,
 		Password: r.Password,
 	}