+++ src/bosh-alicloud-cpi/integration/integration_suite_test.go
@@ -31,7 +31,6 @@ var configForIntegration = string(`{
             "alicloud": {
                 "region_id": "${CPI_REGION}",
 				"zone_id": "${CPI_ZONE}",
-				"endpoint": "${CPI_ENDPOINT}",
                 "access_key_id": "${CPI_ACCESS_KEY_ID}",
                 "access_key_secret": "${CPI_ACCESS_KEY_SECRET}"
             },
@@ -67,12 +66,13 @@ var _ = BeforeSuite(func() {
 
 	logger := boshlog.NewWriterLogger(boshlog.LevelDebug, os.Stderr)
 
-	services := action.Services {
-		Stemcells: alicloud.NewStemcellManager(config),
+	services := action.Services{
+		Stemcells: alicloud.NewStemcellManager(config, logger),
+		Osses:     alicloud.NewOssManager(config, logger),
 		Instances: alicloud.NewInstanceManager(config, logger),
-		Disks: alicloud.NewDiskManager(config, logger),
-		Networks: alicloud.NewNetworkManager(config, logger),
-		Registry: mock.NewRegistryMock(),
+		Disks:     alicloud.NewDiskManager(config, logger),
+		Networks:  alicloud.NewNetworkManager(config, logger),
+		Registry:  mock.NewRegistryMock(),
 	}
 
 	caller = action.NewCallerWithServices(config, logger, services)